v8

+++ src/main/java/com/additional_project/salary_calculator/service/SalaryService.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Service;
 
 @Service
 public interface SalaryService {
-    int calculate(ItemsToCalculate itemsToCalculate);
+    double calculate(ItemsToCalculate itemsToCalculate);
     double getSalary();
 }
